Manage tmux settings through plugins
Turns out a lot of people want the same settings out of tmux. The
tmux-plugin project takes care of a lot of things that were previously
in this file around pane navigation and defaults.

New to these files is the preservation of tmux sessions between
restarts. This has broken a couple times already, but the good outweighs
the bad.